requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.32
httpx==0.27.0
openai==1.37.1
pinecone-client==4.1.2
python-frontmatter==1.1.0
markdown-it-py==3.0.0
numpy==1.26.4
